Evoars_local/Evoars-main/requirements.txt
deep-translator==1.11.4
paddleocr==2.7.3
paddlepaddle>=2.6.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
textwrap3==0.9.2
flask==3.1.0
opencv-python-headless==4.6.0.66
openai-whisper==20240930
moviepy==2.0.0.dev2
ffmpeg-python==0.2.0
soundfile==0.13.1
# TTS==0.22.0
transformers==4.35.2
pydub==0.25.1
# numba==0.60.0
numpy<2.0.0
psutil==5.9.8
arabic-reshaper==3.0.0
python-bidi==0.4.2